rename base query logic file
